pyproject.toml
[tool.poetry]
name = "public-travel-notifications-server"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
#packages = [{include = "public_travel_notifications_server"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.95.0"
uvicorn = "^0.21.1"
firebase-admin = "^6.1.0"
mysql-connector-python = "^8.0.32"
selenium = "^4.8.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
